Use a little caching to improve performance with lots of deleted objects
In case there are a lot of deleted flavors and/or images, horizon pulls
the information from nova/glance several times for every instance. This
causes high load on respective services and very long loading times in
Horizon. Per-request caching of flavors and Django-caching of image
names improves the situation quite a lot. At least in our case.

YMMV of course.

Change-Id: Idafe67f531eee435053bc3d004ab87c240cff6a1